moving some utils to agent_utils.py
